Skip category when error occur when parsing
